Create .gitattributes file for language specific diff matching
Specifying a gitattributes file allows git to use built in matchers that
are language specific to improve diff output. For example, so that it
can give the name of the method the diff is in as a part of the "chunk"
header".

See this blog post for more details: https://tekin.co.uk/2020/10/better-git-diff-output-for-ruby-python-elixir-and-more
Here is the gist this was pulled from: https://gist.github.com/tekin/12500956bd56784728e490d8cef9cb81