Add SQL linter and stop automatically removing ruby whitespace
I now typically remove white space on demand using another command,
which is helpful in codebases where there is a LOT of trailing white
space. It gives me the option of if I want to clean up that file right
then or not depending on if it would create too busy of a diff.